Progresses towards a goal of extracting some of the logic for GPS retrieval into a separate entity and then place presentation details into seprate components.
The idea is to have the retrieval method to be provided methods for
handling the results instead of having the actionCreator methods being
called from within the `then` clause of the Promise.  However, Unit
Tests are unable to be ran successfully because the fetch is a promise
and it's running in parallel (concurrently, actually) instead of waiting
forcing me to wait for the results.  As it's supposed to because it's a
promise.